third_party/blink/web_tests/platform/generic/fast/frames/unload-reparent-sibling-frame-expected.txt

Reparented sibling frames from unload handlers should load.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS frame1.contentDocument is not null
PASS frame2.contentDocument is not null
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: 'frame1'
--------
frame1

--------
Frame: 'frame2'
--------
frame2
